command/cliconfig: New CLI config implementation
This supports both new and old HCL parsers for now, with the intent of
dropping the old parser in a future major release.

The motivation for doing this right now is so that in a future commit we
can use new HCL's "hclwrite" package to do in-place updates on the CLI
configuration files for commands like "terraform login".